Refactor add out into helper files
